refactor: split index tests up into separate files.
